// frontend/src/shared/tour-flight-card/TourFlightCard.types.ts
export interface ITourFlightCard {
  flights: {
    id: number;
    flight_number: string;
    airline: string;
    departure_airport: string;
    arrival_airport: string;
    departure_date: string;
    departure_time: string;
    arrival_date: string;
    arrival_time: string;
    price: string;
    service_class: string;
    flight_type: string;
    description: string;
  }[];
}